pyproject.toml
[project]
name = "livraria"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "carloshenriqueok", email = "ifcarloshenrique@gmail.com"},
]
dependencies = ["django>=5.2.1", "djangorestframework>=3.16.0", "django-cors-headers>=4.7.0", "gunicorn>=23.0.0", "whitenoise>=6.9.0"]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
